Fix: added worker thread so app doesnt freeze when long videos are transcribed (tested on 10min video)
